requirements.txt
videodb
flask
python-dotenv
opencv-python
moviepy
requests
